refactor: group UI scripts
